Extract current limiting strategy into an interface + implementing class
