+++ src/main/java/fi/tkgwf/ruuvi/strategy/LimitingStrategy.java
@@ -1,0 +1,20 @@
+package fi.tkgwf.ruuvi.strategy;
+
+import fi.tkgwf.ruuvi.bean.RuuviMeasurement;
+
+import java.util.Optional;
+
+public interface LimitingStrategy {
+    /**
+     * Applies a limiting strategy to the given measurement.
+     *
+     * @param measurement A measurement considered for persisting.
+     * @return An {@link Optional#empty()} if the strategy is to not store this measurement,
+     * or a non-empty {@link Optional} item to be stored.
+     *
+     * Note that depending on the strategy the returned {@link RuuviMeasurement} may or may not be
+     * the same object that was given in as a parameter: it might be the strategy to, say, calculate
+     * some averages or maximum values of some of the previous measurements instead.
+     */
+    Optional<RuuviMeasurement> apply(RuuviMeasurement measurement);
+}
